+++ lib/src/neu/parsers/parser.ts
@@ -1,50 +1,20 @@
 import * as path from "path";
-import {
-  ClassDeclaration,
-  CompilerOptions,
-  MethodDeclaration,
-  ParameterDeclaration,
-  Project,
-  PropertyDeclaration,
-  SourceFile,
-  ts,
-  TypeGuards
-} from "ts-morph";
+import { CompilerOptions, Project, SourceFile, ts } from "ts-morph";
 import { ApiConfig } from "../../syntax/api";
-import { EndpointConfig } from "../../syntax/endpoint";
-import { ResponseConfig } from "../../syntax/response";
-import {
-  Body,
-  Contract,
-  DefaultResponse,
-  Endpoint,
-  Header,
-  PathParam,
-  QueryParam,
-  Request,
-  Response,
-  Security,
-  SecurityHeader
-} from "../definitions";
+import { Contract, Endpoint, Security } from "../definitions";
 import { LociTable } from "../locations";
 import {
   getClassWithDecoratorOrThrow,
   getDecoratorConfigOrThrow,
   getJsDoc,
-  getMethodWithDecorator,
-  getObjLiteralProp,
   getObjLiteralPropOrThrow,
-  getParameterTypeAsTypeLiteralOrThrow,
-  getParamWithDecorator,
-  getPropertyName,
   getPropertyWithDecorator,
-  getPropValueAsArrayOrThrow,
-  getPropValueAsNumberOrThrow,
   getPropValueAsStringOrThrow,
-  getSelfAndLocalDependencies,
-  isHttpMethod
+  getSelfAndLocalDependencies
 } from "../parser-helpers";
-import { parseType, TypeTable } from "../type-parser";
+import { TypeTable } from "../types";
+import { parseEndpoint } from "./endpoint-parser";
+import { parseSecurityHeader } from "./security-header-parser";
 
 export function parse(sourcePath: string): Contract {
   const project = createProject();
@@ -117,312 +87,6 @@ function parseRootSourceFile(file: SourceFile): Contract {
     types: typeTable.toArray(),
     security,
     endpoints
-  };
-}
-
-function parseSecurityHeader(
-  property: PropertyDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): SecurityHeader {
-  if (property.hasQuestionToken()) {
-    throw new Error("@securityHeader property cannot be optional");
-  }
-  const decorator = property.getDecoratorOrThrow("securityHeader");
-  const name = getPropertyName(property);
-  const descriptionDoc = getJsDoc(property);
-
-  return {
-    name,
-    description: descriptionDoc && descriptionDoc.getComment(),
-    type: parseType(property.getTypeNodeOrThrow(), typeTable, lociTable)
-  };
-}
-
-function parseEndpoint(
-  klass: ClassDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): Endpoint {
-  const decorator = klass.getDecoratorOrThrow("endpoint");
-  const decoratorConfig = getDecoratorConfigOrThrow(decorator);
-  const endpointName = klass.getNameOrThrow();
-  const methodProp = getObjLiteralPropOrThrow<EndpointConfig>(
-    decoratorConfig,
-    "method"
-  );
-  const methodLiteral = getPropValueAsStringOrThrow(methodProp);
-  const methodValue = methodLiteral.getLiteralText();
-  if (!isHttpMethod(methodValue)) {
-    throw new Error(`expected a HttpMethod, got ${methodValue}`);
-  }
-  const pathProp = getObjLiteralPropOrThrow<EndpointConfig>(
-    decoratorConfig,
-    "path"
-  );
-  const pathLiteral = getPropValueAsStringOrThrow(pathProp);
-  const tagsProp = getObjLiteralProp<EndpointConfig>(decoratorConfig, "tags");
-  const tagsLiteral = tagsProp && getPropValueAsArrayOrThrow(tagsProp);
-  const tagsValueLiterals = tagsLiteral
-    ? tagsLiteral.getElements().map(elementExpr => {
-        if (TypeGuards.isStringLiteral(elementExpr)) {
-          return elementExpr;
-        } else {
-          throw new Error(`tag must be a string`);
-        }
-      })
-    : [];
-
-  const descriptionDoc = getJsDoc(klass);
-
-  const requestMethod = getMethodWithDecorator(klass, "request");
-  const request =
-    requestMethod &&
-    parseRequest(requestMethod, typeTable, lociTable, { endpointName });
-
-  const responseMethods = klass
-    .getMethods()
-    .filter(m => m.getDecorator("response") !== undefined);
-
-  const responses = responseMethods
-    .map(r => parseResponse(r, typeTable, lociTable))
-    .sort((a, b) => {
-      if (a.status === b.status) {
-        throw new Error(
-          `Multiple responses found for status code ${a.status} in endpoint ${endpointName}`
-        );
-      }
-      return b.status > a.status ? -1 : 1;
-    });
-
-  const defaultResponseMethod = getMethodWithDecorator(
-    klass,
-    "defaultResponse"
-  );
-
-  const defaultResponse =
-    defaultResponseMethod &&
-    parseDefaultResponse(defaultResponseMethod, typeTable, lociTable);
-
-  // Add location data
-  lociTable.addMorphNode(LociTable.endpointClassKey(endpointName), klass);
-  lociTable.addMorphNode(
-    LociTable.endpointDecoratorKey(endpointName),
-    decorator
-  );
-  lociTable.addMorphNode(LociTable.endpointMethodKey(endpointName), methodProp);
-  lociTable.addMorphNode(LociTable.endpointPathKey(endpointName), pathProp);
-  if (descriptionDoc) {
-    lociTable.addMorphNode(
-      LociTable.endpointDescriptionKey(endpointName),
-      descriptionDoc
-    );
-  }
-  if (tagsProp) {
-    lociTable.addMorphNode(LociTable.endpointTagsKey(endpointName), tagsProp);
-    tagsValueLiterals.forEach(tagLiteral => {
-      lociTable.addMorphNode(
-        LociTable.endpointTagKey(endpointName, tagLiteral.getLiteralText()),
-        tagLiteral
-      );
-    });
-  }
-
-  return {
-    name: endpointName,
-    description: descriptionDoc && descriptionDoc.getComment(),
-    tags: tagsValueLiterals.map(literal => literal.getLiteralText()),
-    method: methodValue,
-    path: pathLiteral.getLiteralText(),
-    request,
-    responses,
-    defaultResponse
-  };
-}
-
-function parseRequest(
-  method: MethodDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable,
-  lociContext: {
-    endpointName: string;
-  }
-): Request {
-  const decorator = method.getDecoratorOrThrow("request");
-  const headersParam = getParamWithDecorator(method, "headers");
-  const pathParamsParam = getParamWithDecorator(method, "pathParams");
-  const queryParamsParam = getParamWithDecorator(method, "queryParams");
-  const bodyParam = getParamWithDecorator(method, "body");
-
-  const headers = headersParam
-    ? parseHeaders(headersParam, typeTable, lociTable)
-    : [];
-
-  const pathParams = pathParamsParam
-    ? parsePathParams(pathParamsParam, typeTable, lociTable)
-    : [];
-
-  const queryParams = queryParamsParam
-    ? parseQueryParams(queryParamsParam, typeTable, lociTable)
-    : [];
-
-  const body = bodyParam && parseBody(bodyParam, typeTable, lociTable);
-
-  // TODO: add loci information
-
-  return {
-    headers,
-    pathParams,
-    queryParams,
-    body
-  };
-}
-
-function parseResponse(
-  method: MethodDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): Response {
-  const decorator = method.getDecoratorOrThrow("response");
-  const decoratorConfig = getDecoratorConfigOrThrow(decorator);
-  const statusProp = getObjLiteralPropOrThrow<ResponseConfig>(
-    decoratorConfig,
-    "status"
-  );
-  const statusLiteral = getPropValueAsNumberOrThrow(statusProp);
-  const headersParam = getParamWithDecorator(method, "headers");
-  const headers = headersParam
-    ? parseHeaders(headersParam, typeTable, lociTable)
-    : [];
-  const bodyParam = getParamWithDecorator(method, "body");
-  const body = bodyParam && parseBody(bodyParam, typeTable, lociTable);
-  const descriptionDoc = getJsDoc(method);
-
-  return {
-    status: statusLiteral.getLiteralValue(),
-    headers,
-    description: descriptionDoc && descriptionDoc.getComment(),
-    body
-  };
-}
-
-function parseDefaultResponse(
-  method: MethodDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): DefaultResponse {
-  const decorator = method.getDecoratorOrThrow("defaultResponse");
-  const headersParam = getParamWithDecorator(method, "headers");
-  const headers = headersParam
-    ? parseHeaders(headersParam, typeTable, lociTable)
-    : [];
-  const bodyParam = getParamWithDecorator(method, "body");
-  const body = bodyParam && parseBody(bodyParam, typeTable, lociTable);
-  const descriptionDoc = getJsDoc(method);
-
-  return {
-    headers,
-    description: descriptionDoc && descriptionDoc.getComment(),
-    body
-  };
-}
-
-function parseHeaders(
-  parameter: ParameterDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): Header[] {
-  const decorator = parameter.getDecoratorOrThrow("headers");
-  // TODO check parameter.isOptional()
-  if (parameter.hasQuestionToken()) {
-    throw new Error("@headers parameter cannot be optional");
-  }
-  const type = getParameterTypeAsTypeLiteralOrThrow(parameter);
-  const headers = type
-    .getProperties()
-    .map(p => {
-      const pDescription = getJsDoc(p);
-      return {
-        name: getPropertyName(p),
-        type: parseType(p.getTypeNodeOrThrow(), typeTable, lociTable),
-        description: pDescription && pDescription.getComment(),
-        optional: p.hasQuestionToken()
-      };
-    })
-    .sort((a, b) => (b.name > a.name ? -1 : 1));
-  // TODO: add loci information
-  return headers;
-}
-
-function parsePathParams(
-  parameter: ParameterDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): PathParam[] {
-  const decorator = parameter.getDecoratorOrThrow("pathParams");
-  if (parameter.hasQuestionToken()) {
-    throw new Error("@pathParams parameter cannot be optional");
-  }
-  const type = getParameterTypeAsTypeLiteralOrThrow(parameter);
-  const pathParams = type
-    .getProperties()
-    .map(p => {
-      const pDescription = getJsDoc(p);
-      if (parameter.hasQuestionToken()) {
-        throw new Error("@pathParams properties cannot be optional");
-      }
-      return {
-        name: getPropertyName(p),
-        type: parseType(p.getTypeNodeOrThrow(), typeTable, lociTable),
-        description: pDescription && pDescription.getComment()
-      };
-    })
-    .sort((a, b) => (b.name > a.name ? -1 : 1));
-  // TODO: add loci information
-  return pathParams;
-}
-
-function parseQueryParams(
-  parameter: ParameterDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): QueryParam[] {
-  const decorator = parameter.getDecoratorOrThrow("queryParams");
-  if (parameter.hasQuestionToken()) {
-    throw new Error("@queryParams parameter cannot be optional");
-  }
-  const type = getParameterTypeAsTypeLiteralOrThrow(parameter);
-  const queryParams = type
-    .getProperties()
-    .map(p => {
-      const pDescription = getJsDoc(p);
-      return {
-        name: getPropertyName(p),
-        type: parseType(p.getTypeNodeOrThrow(), typeTable, lociTable),
-        description: pDescription && pDescription.getComment(),
-        optional: p.hasQuestionToken()
-      };
-    })
-    .sort((a, b) => (b.name > a.name ? -1 : 1));
-  // TODO: add loci information
-  return queryParams;
-}
-
-function parseBody(
-  parameter: ParameterDeclaration,
-  typeTable: TypeTable,
-  lociTable: LociTable
-): Body {
-  const decorator = parameter.getDecoratorOrThrow("body");
-  if (parameter.hasQuestionToken()) {
-    throw new Error("@body parameter cannot be optional");
-  }
-  const type = parseType(parameter.getTypeNodeOrThrow(), typeTable, lociTable);
-  // TODO: add loci information
-  return {
-    // TODO: how to extract description from parameter declaration?
-    description: undefined,
-    type
   };
 }
